refactor(frontend): the Settings tab out of the panel, with its three copies made one
`panel-settings.ts` takes the rail and phone index, the autosaving option cards,
the Profiles and Notifications editors and the Companions list. `_makeForm` and
`_patchSettingsSection` stay on the panel — the first because being registered in
`_liveHassEls` is what keeps a form's pickers current, the second because it is a
render-path optimization `_applyLocation` reaches for.

The region arrived carrying four sets of duplicates, each now written once:

- The collapsible section card (header, count, remembered collapse, intro) that
  Profiles and Notifications were each building inline: `settingsListSection`.
- The collapsible row a profile, a notification and a profile's sync group were
  each building inline — header, chevron, remembered expansion, Delete: `itemCard`.
- The optimistic save both option lists ran through, twice over:
  `persistOptionList`, with the options key and the debounce key threaded as one.
- Four hand-rolled `ha-form` blocks reimplementing `_makeForm`. `_makeForm` now
  takes an optional *labelling*, because those four label their fields from four
  different key spaces (`settings.`, `notify.`, `todo_sync.`) — and a *labelling*
  without a `computeHelper` leaves the form without one, rather than falling back
  to the `help.<name>` lookup and putting prose under a field that never had any.

Counted at the source: five `createElement('ha-form')` become one, three item
scaffolds one, two section scaffolds one, two persist runs one. The deduplicated
set goes from 358 to 312 lines of code.

Emitted HTML is unchanged. Every template literal in the built bundle is
byte-identical to the one before; the two that are gone are exactly the second
copy of the section header, which is the duplicate this removes.
